refactor(anthropic): deduplicate sequencing, media-type, and test helpers
- Move hashFile from files.go to files_test.go (only used in tests)
- Remove marshalToMapBeta/contentArrayBeta aliases; use originals directly
- Extract generic validateSequencing[T] and repairSequencing[T] to unify
  standard and beta sequencing validation/repair logic
- Replace duplicated inline media-type switch with extractMediaType in
  both convertUserMultiContent and convertBetaUserMultiContent
